@@ -1,8 +1,10 @@
 package config
 
+// CfgFunc - конфигурационная функция для Compose.
+// Принимает на вход текущую конфигурацию и возвращает новую конфигурацию.
 type CfgFunc func(*Config) (*Config, error)
 
-// Compose - объединяет набор конфигурационных функций в одну
+// Compose - объединяет набор конфигурационных функций в итоговую конфигурацию
 func Compose(funcs ...CfgFunc) (*Config, error) {
 	var err error
 	cfg := &Config{}
