@@ -3,13 +3,14 @@ package config
 import (
 	"flag"
 	"fmt"
-	"github.com/caarlos0/env/v6"
-	"golang.org/x/sync/errgroup"
 	"net"
 	"net/url"
 	"os"
 	"strings"
 	"time"
+
+	"github.com/caarlos0/env/v6"
+	"golang.org/x/sync/errgroup"
 )
 
 // Config - конфигурация приложения
@@ -35,7 +36,7 @@ type Config struct {
 }
 
 // Default - конфигурационная функция, которая возвращает конфигурацию по умолчанию.
-// Входной параметр не используется. Ошибки не возвращаются.
+// Входной параметр не используется.
 func Default(_ *Config) (*Config, error) {
 	secret, err := randSecret(64)
 	if err != nil {
@@ -57,11 +58,12 @@ func Default(_ *Config) (*Config, error) {
 // FromCLI - конфигурационная функция, которая считывает конфигурацию приложения из переменных окружения.
 //
 // Флаги командной строки:
-//    -a <host:port> - адрес для запуска HTTP-сервера
-//    -b <url>       - базовый адрес сокращённого URL
-//    -f <path>      - файл для хранения данных
-//    -t <duration>  - время жизни авторизационного токена
-//	  -d <dsn>       - строка с адресом подключения к БД
+//
+//	-a <host:port> - адрес для запуска HTTP-сервера
+//	-b <url>       - базовый адрес сокращённого URL
+//	-f <path>      - файл для хранения данных
+//	-t <duration>  - время жизни авторизационного токена
+//	-d <dsn>       - строка с адресом подключения к БД
 //
 // Если какие-либо значения не заданы в командной строке, то используются значения переданные в cfg.
 func FromCLI(cfg *Config) (*Config, error) {
@@ -91,11 +93,12 @@ func fromCLI(cfg *Config, arguments ...string) (*Config, error) {
 // FromEnv - конфигурационная функция, которая читывает конфигурацию приложения из переменных окружения.
 //
 // Переменные окружения:
-//    SERVER_ADDRESS    - адрес для запуска HTTP-сервера
-//    BASE_URL          - базовый адрес сокращённого URL
-//    FILE_STORAGE_PATH - файл для хранения данных
-//    AUTH_TTL          - время жизни авторизационного токена
-//	  AUTH_SECRET       - секретный ключ для подписи авторизационного токена
+//
+//	SERVER_ADDRESS    - адрес для запуска HTTP-сервера
+//	BASE_URL          - базовый адрес сокращённого URL
+//	FILE_STORAGE_PATH - файл для хранения данных
+//	AUTH_TTL          - время жизни авторизационного токена
+//	AUTH_SECRET       - секретный ключ для подписи авторизационного токена
 //
 // Если какие-либо переменные окружения не заданы, то используются значения переданные в cfg.
 func FromEnv(cfg *Config) (*Config, error) {
@@ -121,8 +124,9 @@ func (c *Config) validate() error {
 
 // validateBaseURL - проверяет базовый адрес сокращённого URL.
 // Возвращает ошибку в случае:
-//    - URL не содержит протокол http или https
-//    - URL содержит параметры или фрагмент.
+//   - URL не содержит протокол http или https
+//   - URL содержит параметры или фрагмент.
+//
 // Добавляет слеш в конце Path, если его нет.
 func (c *Config) validateBaseURL() error {
 	if c.BaseURL.RawQuery != "" || c.BaseURL.Fragment != "" {
