@@ -3,22 +3,37 @@ package handlers
 import (
 	"encoding/json"
 	"errors"
-	"github.com/go-chi/chi/v5"
-	"github.com/ofstudio/go-shortener/internal/app/services"
-	"github.com/ofstudio/go-shortener/internal/middleware"
 	"mime"
 	"net/http"
+
+	"github.com/go-chi/chi/v5"
+
+	"github.com/ofstudio/go-shortener/internal/app/services"
+	"github.com/ofstudio/go-shortener/internal/middleware"
 )
+
+// @Title Go-Shortener API
+// @Description API для сокращения ссылок
+// @Version 1.0
+// @Contact.name Oleg Fomin
+// @Contact.email ofstudio@yandex.ru
+// @BasePath /api
+
+// @securityDefinitions.apikey ApiKeyAuth
+// @In cookie
+// @Name auth_token
 
 // APIHandlers - HTTP-хендлеры для JSON API
 type APIHandlers struct {
 	srv *services.Container
 }
 
+// NewAPIHandlers - конструктор APIHandlers
 func NewAPIHandlers(srv *services.Container) *APIHandlers {
 	return &APIHandlers{srv}
 }
 
+// Routes - возвращает роутер с хендлерами
 func (h APIHandlers) Routes() chi.Router {
 	r := chi.NewRouter()
 	r.Post("/shorten", h.shortURLCreate)
@@ -29,10 +44,26 @@ func (h APIHandlers) Routes() chi.Router {
 }
 
 // shortURLCreate - принимает в теле запроса строку URL для сокращения:
-//    {"url":"<url>"}
+//
+//	{"url":"<url>"}
 //
 // Возвращает ответ http.StatusCreated (201) и сокращенный URL в виде JSON:
-//    {"result":"<shorten_url>"}
+//
+//	{"result":"<shorten_url>"}
+//
+// @Tags shorten
+// @Summary Создает сокращенную ссылку
+// @Security cookieAuth
+// @ID shortURLCreate
+// @Accept  json
+// @Produce json
+// @Param   request body handlers.shortURLCreate.reqType true "Запрос"
+// @Success 201 {object} handlers.shortURLCreate.resType
+// @Failure 400
+// @Failure 401
+// @Failure 409 {object} handlers.shortURLCreate.resType
+// @Failure 500
+// @Router /shorten [post]
 func (h APIHandlers) shortURLCreate(w http.ResponseWriter, r *http.Request) {
 	// Структура запроса
 	type reqType struct {
@@ -79,21 +110,37 @@ func (h APIHandlers) shortURLCreate(w http.ResponseWriter, r *http.Request) {
 }
 
 // shortURLCreateBatch - принимает в теле запроса список строк URL для сокращения:
-//    [
-//        {
-//            "correlation_id": "<строковый идентификатор>",
-//            "original_url": "<URL для сокращения>"
-//        },
-//        ...
-//    ]
+//
+//	[
+//	    {
+//	        "correlation_id": "<строковый идентификатор>",
+//	        "original_url": "<URL для сокращения>"
+//	    },
+//	    ...
+//	]
+//
 // Возвращает ответ http.StatusCreated (201) и сокращенный URL в виде JSON:
-//    [
-//        {
-//            "correlation_id": "<строковый идентификатор из объекта запроса>",
-//            "short_url": "<shorten_url>"
-//        },
-//        ...
-//    ]
+//
+//	[
+//	    {
+//	        "correlation_id": "<строковый идентификатор из объекта запроса>",
+//	        "short_url": "<shorten_url>"
+//	    },
+//	    ...
+//	]
+//
+// @Tags shorten
+// @Summary Создает несколько сокращенных ссылок
+// @Security cookieAuth
+// @ID shortURLCreateBatch
+// @Accept  json
+// @Produce json
+// @Param   request body handlers.shortURLCreateBatch.reqType true "Запрос"
+// @Success 201 {array} handlers.shortURLCreateBatch.resType
+// @Failure 400 {string} Bad Request
+// @Failure 401 {string} string "Unauthorized"
+// @Failure 500 {string} string "Internal Server Error"
+// @Router /shorten/batch [post]
 func (h APIHandlers) shortURLCreateBatch(w http.ResponseWriter, r *http.Request) {
 	// Структура элемента запроса
 	type reqType struct {
@@ -140,8 +187,23 @@ func (h APIHandlers) shortURLCreateBatch(w http.ResponseWriter, r *http.Request)
 
 // shortURLDeleteBatch - помечает ссылки пользователя как удаленные.
 // Формат запроса:
-//    [ "a", "b", "c", "d", ...]
+//
+//	[ "a", "b", "c", "d", ...]
+//
 // Возвращает ответ http.StatusAccepted (202)
+//
+// @Tags user
+// @Summary Удаляет несколько сокращенных ссылок
+// @Security cookieAuth
+// @ID shortURLDeleteBatch
+// @Accept  json
+// @Produce json
+// @Param   request body handlers.shortURLDeleteBatch.reqType true "Запрос"
+// @Success 202
+// @Failure 400
+// @Failure 401
+// @Failure 500
+// @Router /user/urls [delete]
 func (h APIHandlers) shortURLDeleteBatch(w http.ResponseWriter, r *http.Request) {
 	// Структура элемента запроса
 	type reqType []string
@@ -173,13 +235,25 @@ func (h APIHandlers) shortURLDeleteBatch(w http.ResponseWriter, r *http.Request)
 
 // shortURLGetByUserID - возвращает список сокращенных ссылок пользователя.
 // Формат ответа:
-//    [
-//        {
-//            "short_url": "http://...",
-//            "original_url": "http://..."
-//        },
-//        ...
-//    ]
+//
+//	[
+//	    {
+//	        "short_url": "http://...",
+//	        "original_url": "http://..."
+//	    },
+//	    ...
+//	]
+//
+// @Tags user
+// @Summary Возвращает список сокращенных ссылок пользователя
+// @Security cookieAuth
+// @ID shortURLGetByUserID
+// @Produce json
+// @Success 200 {array} handlers.shortURLGetByUserID.resType
+// @Failure 400
+// @Failure 401
+// @Failure 500
+// @Router /user/urls [get]
 func (h APIHandlers) shortURLGetByUserID(w http.ResponseWriter, r *http.Request) {
 	// Структура ответа
 	type resType struct {
