@@ -1,9 +1,10 @@
 package middleware_test
 
 import (
+	"testing"
+
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"testing"
 )
 
 func TestCompressorSuite(t *testing.T) {
