@@ -1,15 +1,16 @@
 package repo
 
 import (
-	"github.com/ofstudio/go-shortener/internal/app/config"
 	"log"
+
+	"github.com/ofstudio/go-shortener/internal/app/config"
 )
 
 // Fabric - фабрика для создания репозитория.
 // Если задан DatabaseDSN - используем SQL-репозиторий.
 // Иначе, если задан fileStoragePath — используем AOF-репозиторий.
 // Иначе используем репозиторий в памяти.
-func Fabric(cfg *config.Config) (Repo, error) {
+func Fabric(cfg *config.Config) (IRepo, error) {
 	switch {
 	case cfg.DatabaseDSN != "":
 		log.Print("Using Postgres storage")
